requirements.txt
streamlit==1.37.1
streamlit-oauth==0.1.13
weaviate-client==4.7.1
openai==1.40.8
plotly==5.24.0
pydantic==2.8.2
stqdm==0.0.5